requirements.txt
Django
djangorestframework
Pillow
pytz
